pyproject.toml
[tool.poetry]
name = "climate-disaster-predictor"
version = "0.1.0"
description = "y"
authors = ["Kevin Weingarten <weingarten.kevin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
pandas = "^1.5.1"
fastparquet = "^0.8.3"
beautifulsoup4 = "^4.9.0"
SQLAlchemy = "^1.4.43"
psycopg2 = "^2.9.5"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
